requirements.txt
asgiref==3.7.2
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
lxml==4.9.3
pdf2image==1.16.3
Pillow==10.1.0
PyJWT==2.8.0
python-docx==1.0.1
python-magic==0.4.27
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
